pyproject.toml
[project]
name = "gitnema"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "pillow>=12.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
